+++ src/lib.rs
@@ -55,6 +55,12 @@
 
 #![cfg_attr(not(test), no_std)]
 #![deny(missing_docs)]
+#![cfg_attr(feature = "asynch", feature(async_fn_traits))]
+// Since we don't have MSRV we always enable this feature when `asynch` is enabled
+#![cfg_attr(feature = "asynch", feature(async_fn_in_trait))]
+#![cfg_attr(feature = "asynch", allow(stable_features))]
+#![cfg_attr(feature = "asynch", allow(async_fn_in_trait))]
+#![cfg_attr(feature = "asynch", feature(async_closure))]
 
 // ****************************************************************************
 //