+++ src/sdcard/mod.rs
@@ -8,7 +8,7 @@
 pub mod proto;
 
 use crate::{trace, Block, BlockCount, BlockDevice, BlockIdx};
-use core::cell::RefCell;
+use core::{cell::RefCell, marker::PhantomData, ops::AsyncFnOnce};
 use proto::*;
 
 // ****************************************************************************
@@ -56,6 +56,24 @@ impl embedded_hal::digital::OutputPin for DummyCsPin {
     }
 }
 
+/// Blocking SD card API mode and Delay
+pub struct Blocking;
+impl Mode for Blocking {}
+
+/// async/await SD card API mode and Delay
+pub struct Async;
+impl Mode for Async {}
+
+#[allow(private_bounds)]
+/// Async or Blocking mode of the APIs
+pub trait Mode: sealed::SealedMode {}
+
+mod sealed {
+    impl SealedMode for super::Blocking {}
+    impl SealedMode for super::Async {}
+    pub(super) trait SealedMode {}
+}
+
 /// Represents an SD Card on an SPI bus.
 ///
 /// Built from an [`SpiDevice`] implementation and a Chip Select pin.
@@ -69,16 +87,70 @@ impl embedded_hal::digital::OutputPin for DummyCsPin {
 /// All the APIs take `&self` - mutability is handled using an inner `RefCell`.
 ///
 /// [`SpiDevice`]: embedded_hal::spi::SpiDevice
-pub struct SdCard<SPI, CS, DELAYER>
-where
-    SPI: embedded_hal::spi::SpiDevice<u8>,
-    CS: embedded_hal::digital::OutputPin,
-    DELAYER: embedded_hal::delay::DelayNs,
-{
-    inner: RefCell<SdCardInner<SPI, CS, DELAYER>>,
+pub struct SdCard<SPI, CS, DELAYER, M: Mode> {
+    inner: RefCell<SdCardInner<SPI, CS, DELAYER, M>>,
 }
 
-impl<SPI, CS, DELAYER> SdCard<SPI, CS, DELAYER>
+impl<SPI, CS, DELAYER, M> SdCard<SPI, CS, DELAYER, M>
+where
+    M: Mode,
+{
+    /// Construct a new SD/MMC Card driver, using a raw SPI interface and the given options.
+    ///
+    /// See the docs of the [`SdCard`] struct for more information about
+    /// how to construct the needed `SPI` and `CS` types.
+    ///
+    /// The card will not be initialised at this time. Initialisation is
+    /// deferred until a method is called on the object.
+    pub fn new_with_options(
+        spi: SPI,
+        cs: CS,
+        delayer: DELAYER,
+        options: AcquireOpts,
+    ) -> SdCard<SPI, CS, DELAYER, M> {
+        SdCard {
+            inner: RefCell::new(SdCardInner::<_, _, _, M> {
+                spi,
+                cs,
+                delayer,
+                card_type: None,
+                options,
+                mode: Default::default(),
+            }),
+        }
+    }
+
+    /// Tell the driver the card has been initialised.
+    ///
+    /// This is here in case you were previously using the SD Card, and then a
+    /// previous instance of this object got destroyed but you know for certain
+    /// the SD Card remained powered up and initialised, and you'd just like to
+    /// read/write to/from the card again without going through the
+    /// initialisation sequence again.
+    ///
+    /// # Safety
+    ///
+    /// Only do this if the SD Card has actually been initialised. That is, if
+    /// you have been through the card initialisation sequence as specified in
+    /// the SD Card Specification by sending each appropriate command in turn,
+    /// either manually or using another variable of this [`SdCard`]. The card
+    /// must also be of the indicated type. Failure to uphold this will cause
+    /// data corruption.
+    pub unsafe fn mark_card_as_init(&self, card_type: CardType) {
+        let mut inner = self.inner.borrow_mut();
+        inner.card_type = Some(card_type);
+    }
+
+    /// Mark the card as requiring a reset.
+    ///
+    /// The next operation will assume the card has been freshly inserted.
+    pub fn mark_card_uninit(&self) {
+        let mut inner = self.inner.borrow_mut();
+        inner.card_type = None;
+    }
+}
+
+impl<SPI, CS, DELAYER> SdCard<SPI, CS, DELAYER, Blocking>
 where
     SPI: embedded_hal::spi::SpiDevice<u8>,
     CS: embedded_hal::digital::OutputPin,
@@ -93,32 +165,13 @@ where
     /// deferred until a method is called on the object.
     ///
     /// Uses the default options.
-    pub fn new(spi: SPI, cs: CS, delayer: DELAYER) -> SdCard<SPI, CS, DELAYER> {
-        Self::new_with_options(spi, cs, delayer, AcquireOpts::default())
-    }
-
-    /// Construct a new SD/MMC Card driver, using a raw SPI interface and the given options.
-    ///
-    /// See the docs of the [`SdCard`] struct for more information about
-    /// how to construct the needed `SPI` and `CS` types.
-    ///
-    /// The card will not be initialised at this time. Initialisation is
-    /// deferred until a method is called on the object.
-    pub fn new_with_options(
-        spi: SPI,
-        cs: CS,
-        delayer: DELAYER,
-        options: AcquireOpts,
-    ) -> SdCard<SPI, CS, DELAYER> {
-        SdCard {
-            inner: RefCell::new(SdCardInner {
-                spi,
-                cs,
-                delayer,
-                card_type: None,
-                options,
-            }),
-        }
+    pub fn new(spi: SPI, cs: CS, delayer: DELAYER) -> SdCard<SPI, CS, DELAYER, Blocking>
+    where
+        SPI: embedded_hal::spi::SpiDevice<u8>,
+        CS: embedded_hal::digital::OutputPin,
+        DELAYER: embedded_hal::delay::DelayNs,
+    {
+        SdCard::new_with_options(spi, cs, delayer, AcquireOpts::default())
     }
 
     /// Get a temporary borrow on the underlying SPI device.
@@ -153,15 +206,6 @@ where
         inner.check_init()?;
         inner.erase_single_block_enabled()
     }
-
-    /// Mark the card as requiring a reset.
-    ///
-    /// The next operation will assume the card has been freshly inserted.
-    pub fn mark_card_uninit(&self) {
-        let mut inner = self.inner.borrow_mut();
-        inner.card_type = None;
-    }
-
     /// Get the card type.
     ///
     /// This will trigger card (re-)initialisation.
@@ -170,30 +214,72 @@ where
         inner.check_init().ok()?;
         inner.card_type
     }
+}
 
-    /// Tell the driver the card has been initialised.
+impl<SPI, CS, DELAYER> SdCard<SPI, CS, DELAYER, Async>
+where
+    SPI: embedded_hal_async::spi::SpiDevice<u8>,
+    CS: embedded_hal::digital::OutputPin,
+    DELAYER: embedded_hal_async::delay::DelayNs,
+{
+
+
+
+    /// Create an async SDCard API
+    #[cfg(feature = "asynch")]
+    pub fn new_async(spi: SPI, cs: CS, delayer: DELAYER) -> SdCard<SPI, CS, DELAYER, Async>
+    where
+        SPI: embedded_hal_async::spi::SpiDevice<u8>,
+        CS: embedded_hal::digital::OutputPin,
+        DELAYER: embedded_hal_async::delay::DelayNs,
+    {
+        SdCard::new_with_options(spi, cs, delayer, AcquireOpts::default())
+    }
+
+    /// Get a temporary borrow on the underlying SPI device.
     ///
-    /// This is here in case you were previously using the SD Card, and then a
-    /// previous instance of this object got destroyed but you know for certain
-    /// the SD Card remained powered up and initialised, and you'd just like to
-    /// read/write to/from the card again without going through the
-    /// initialisation sequence again.
+    /// The given closure will be called exactly once, and will be passed a
+    /// mutable reference to the underlying SPI object.
     ///
-    /// # Safety
-    ///
-    /// Only do this if the SD Card has actually been initialised. That is, if
-    /// you have been through the card initialisation sequence as specified in
-    /// the SD Card Specification by sending each appropriate command in turn,
-    /// either manually or using another variable of this [`SdCard`]. The card
-    /// must also be of the indicated type. Failure to uphold this will cause
-    /// data corruption.
-    pub unsafe fn mark_card_as_init(&self, card_type: CardType) {
+    /// Useful if you need to re-clock the SPI, but does not perform card
+    /// initialisation.
+    pub async fn spi<T, F>(&self, func: F) -> T
+    where
+        F: AsyncFnOnce(&mut SPI) -> T,
+    {
         let mut inner = self.inner.borrow_mut();
-        inner.card_type = Some(card_type);
+        func(&mut inner.spi).await
+    }
+
+    /// Return the usable size of this SD card in bytes.
+    ///
+    /// This will trigger card (re-)initialisation.
+    pub async fn num_bytes(&self) -> Result<u64, Error> {
+        let mut inner = self.inner.borrow_mut();
+        inner.check_init().await?;
+        inner.num_bytes().await
+    }
+
+    /// Can this card erase single blocks?
+    ///
+    /// This will trigger card (re-)initialisation.
+    pub async fn erase_single_block_enabled(&self) -> Result<bool, Error> {
+        let mut inner = self.inner.borrow_mut();
+        inner.check_init().await?;
+        inner.erase_single_block_enabled().await
+    }
+
+    /// Get the card type.
+    ///
+    /// This will trigger card (re-)initialisation.
+    pub async fn get_card_type(&self) -> Option<CardType> {
+        let mut inner = self.inner.borrow_mut();
+        inner.check_init().await.ok()?;
+        inner.card_type
     }
 }
 
-impl<SPI, CS, DELAYER> BlockDevice for SdCard<SPI, CS, DELAYER>
+impl<SPI, CS, DELAYER> BlockDevice for SdCard<SPI, CS, DELAYER, Blocking>
 where
     SPI: embedded_hal::spi::SpiDevice<u8>,
     CS: embedded_hal::digital::OutputPin,
@@ -241,23 +327,68 @@ where
     }
 }
 
+#[cfg(feature = "asynch")]
+impl<SPI, CS, DELAYER> crate::blockdevice::AsyncBlockDevice for SdCard<SPI, CS, DELAYER, Async>
+where
+    SPI: embedded_hal_async::spi::SpiDevice<u8>,
+    CS: embedded_hal::digital::OutputPin,
+    DELAYER: embedded_hal_async::delay::DelayNs,
+{
+    type Error = Error;
+
+    /// Read one or more blocks, starting at the given block index.
+    ///
+    /// This will trigger card (re-)initialisation.
+    async fn read(
+        &self,
+        blocks: &mut [Block],
+        start_block_idx: BlockIdx,
+        _reason: &str,
+    ) -> Result<(), Self::Error> {
+        let mut inner = self.inner.borrow_mut();
+        debug!(
+            "Read {} blocks @ {} for {}",
+            blocks.len(),
+            start_block_idx.0,
+            _reason
+        );
+        inner.check_init().await?;
+        inner.read(blocks, start_block_idx).await
+    }
+
+    /// Write one or more blocks, starting at the given block index.
+    ///
+    /// This will trigger card (re-)initialisation.
+    async fn write(&self, blocks: &[Block], start_block_idx: BlockIdx) -> Result<(), Self::Error> {
+        let mut inner = self.inner.borrow_mut();
+        debug!("Writing {} blocks @ {}", blocks.len(), start_block_idx.0);
+        inner.check_init().await?;
+        inner.write(blocks, start_block_idx).await
+    }
+
+    /// Determine how many blocks this device can hold.
+    ///
+    /// This will trigger card (re-)initialisation.
+    async fn num_blocks(&self) -> Result<BlockCount, Self::Error> {
+        let mut inner = self.inner.borrow_mut();
+        inner.check_init().await?;
+        inner.num_blocks().await
+    }
+}
+
 /// Represents an SD Card on an SPI bus.
 ///
 /// All the APIs required `&mut self`.
-struct SdCardInner<SPI, CS, DELAYER>
-where
-    SPI: embedded_hal::spi::SpiDevice<u8>,
-    CS: embedded_hal::digital::OutputPin,
-    DELAYER: embedded_hal::delay::DelayNs,
-{
+struct SdCardInner<SPI, CS, DELAYER, M: Mode> {
     spi: SPI,
     cs: CS,
     delayer: DELAYER,
     card_type: Option<CardType>,
     options: AcquireOpts,
+    mode: PhantomData<M>,
 }
 
-impl<SPI, CS, DELAYER> SdCardInner<SPI, CS, DELAYER>
+impl<SPI, CS, DELAYER> SdCardInner<SPI, CS, DELAYER, Blocking>
 where
     SPI: embedded_hal::spi::SpiDevice<u8>,
     CS: embedded_hal::digital::OutputPin,
@@ -655,6 +786,425 @@ where
     }
 }
 
+#[cfg(feature = "asynch")]
+impl<SPI, CS, DELAYER> SdCardInner<SPI, CS, DELAYER, Async>
+where
+    SPI: embedded_hal_async::spi::SpiDevice<u8>,
+    CS: embedded_hal::digital::OutputPin,
+    DELAYER: embedded_hal_async::delay::DelayNs,
+{
+    /// Read one or more blocks, starting at the given block index.
+    async fn read(&mut self, blocks: &mut [Block], start_block_idx: BlockIdx) -> Result<(), Error> {
+        let start_idx = match self.card_type {
+            Some(CardType::SD1 | CardType::SD2) => start_block_idx.0 * 512,
+            Some(CardType::SDHC) => start_block_idx.0,
+            None => return Err(Error::CardNotFound),
+        };
+        self.with_chip_select(async |s| {
+            if blocks.len() == 1 {
+                // Start a single-block read
+                s.card_command(CMD17, start_idx).await?;
+                s.read_data(&mut blocks[0].contents).await?;
+            } else {
+                // Start a multi-block read
+                s.card_command(CMD18, start_idx).await?;
+                for block in blocks.iter_mut() {
+                    s.read_data(&mut block.contents).await?;
+                }
+                // Stop the read
+                s.card_command(CMD12, 0).await?;
+            }
+            Ok(())
+        })
+        .await
+    }
+
+    /// Write one or more blocks, starting at the given block index.
+    async fn write(&mut self, blocks: &[Block], start_block_idx: BlockIdx) -> Result<(), Error> {
+        let start_idx = match self.card_type {
+            Some(CardType::SD1 | CardType::SD2) => start_block_idx.0 * 512,
+            Some(CardType::SDHC) => start_block_idx.0,
+            None => return Err(Error::CardNotFound),
+        };
+        self.with_chip_select(async |s| {
+            if blocks.len() == 1 {
+                // Start a single-block write
+                s.card_command(CMD24, start_idx).await?;
+                s.write_data(DATA_START_BLOCK, &blocks[0].contents).await?;
+                s.wait_not_busy(Delay::<Async>::new_write()).await?;
+                if s.card_command(CMD13, 0).await? != 0x00 {
+                    return Err(Error::WriteError);
+                }
+                if s.read_byte().await? != 0x00 {
+                    return Err(Error::WriteError);
+                }
+            } else {
+                // > It is recommended using this command preceding CMD25, some of the cards will be faster for Multiple
+                // > Write Blocks operation. Note that the host should send ACMD23 just before WRITE command if the host
+                // > wants to use the pre-erased feature
+                s.card_acmd(ACMD23, blocks.len() as u32).await?;
+                // wait for card to be ready before sending the next command
+                s.wait_not_busy(Delay::new_write()).await?;
+
+                // Start a multi-block write
+                s.card_command(CMD25, start_idx).await?;
+                for block in blocks.iter() {
+                    s.wait_not_busy(Delay::new_write()).await?;
+                    s.write_data(WRITE_MULTIPLE_TOKEN, &block.contents).await?;
+                }
+                // Stop the write
+                s.wait_not_busy(Delay::new_write()).await?;
+                s.write_byte(STOP_TRAN_TOKEN).await?;
+            }
+            Ok(())
+        })
+        .await
+    }
+
+    /// Determine how many blocks this device can hold.
+    async fn num_blocks(&mut self) -> Result<BlockCount, Error> {
+        let num_blocks = self
+            .with_chip_select(async |s| {
+                let csd = s.read_csd().await?;
+                debug!("CSD: {:?}", csd);
+                match csd {
+                    Csd::V1(ref contents) => Ok(contents.card_capacity_blocks()),
+                    Csd::V2(ref contents) => Ok(contents.card_capacity_blocks()),
+                }
+            })
+            .await?;
+        Ok(BlockCount(num_blocks))
+    }
+
+    /// Return the usable size of this SD card in bytes.
+    async fn num_bytes(&mut self) -> Result<u64, Error> {
+        self.with_chip_select(async |s| {
+            let csd = s.read_csd().await?;
+            debug!("CSD: {:?}", csd);
+            match csd {
+                Csd::V1(ref contents) => Ok(contents.card_capacity_bytes()),
+                Csd::V2(ref contents) => Ok(contents.card_capacity_bytes()),
+            }
+        })
+        .await
+    }
+
+    /// Can this card erase single blocks?
+    pub async fn erase_single_block_enabled(&mut self) -> Result<bool, Error> {
+        self.with_chip_select(async |s| {
+            let csd = s.read_csd().await?;
+            match csd {
+                Csd::V1(ref contents) => Ok(contents.erase_single_block_enabled()),
+                Csd::V2(ref contents) => Ok(contents.erase_single_block_enabled()),
+            }
+        })
+        .await
+    }
+
+    /// Read the 'card specific data' block.
+    async fn read_csd(&mut self) -> Result<Csd, Error> {
+        match self.card_type {
+            Some(CardType::SD1) => {
+                let mut csd = CsdV1::new();
+                if self.card_command(CMD9, 0).await? != 0 {
+                    return Err(Error::RegisterReadError);
+                }
+                self.read_data(&mut csd.data).await?;
+                Ok(Csd::V1(csd))
+            }
+            Some(CardType::SD2 | CardType::SDHC) => {
+                let mut csd = CsdV2::new();
+                if self.card_command(CMD9, 0).await? != 0 {
+                    return Err(Error::RegisterReadError);
+                }
+                self.read_data(&mut csd.data).await?;
+                Ok(Csd::V2(csd))
+            }
+            None => Err(Error::CardNotFound),
+        }
+    }
+
+    /// Read an arbitrary number of bytes from the card using the SD Card
+    /// protocol and an optional CRC. Always fills the given buffer, so make
+    /// sure it's the right size.
+    async fn read_data(&mut self, buffer: &mut [u8]) -> Result<(), Error> {
+        // Get first non-FF byte.
+        let mut delay = Delay::<Async>::new_read();
+        let status = loop {
+            let s = self.read_byte().await?;
+            if s != 0xFF {
+                break s;
+            }
+            delay
+                .delay_async(&mut self.delayer, Error::TimeoutReadBuffer)
+                .await?;
+        };
+        if status != DATA_START_BLOCK {
+            return Err(Error::ReadError);
+        }
+
+        for b in buffer.iter_mut() {
+            *b = 0xFF;
+        }
+        self.transfer_bytes(buffer).await?;
+
+        // These two bytes are always sent. They are either a valid CRC, or
+        // junk, depending on whether CRC mode was enabled.
+        let mut crc_bytes = [0xFF; 2];
+        self.transfer_bytes(&mut crc_bytes).await?;
+        if self.options.use_crc {
+            let crc = u16::from_be_bytes(crc_bytes);
+            let calc_crc = crc16(buffer);
+            if crc != calc_crc {
+                return Err(Error::CrcError(crc, calc_crc));
+            }
+        }
+
+        Ok(())
+    }
+
+    /// Write an arbitrary number of bytes to the card using the SD protocol and
+    /// an optional CRC.
+    async fn write_data(&mut self, token: u8, buffer: &[u8]) -> Result<(), Error> {
+        self.write_byte(token).await?;
+        self.write_bytes(buffer).await?;
+        let crc_bytes = if self.options.use_crc {
+            crc16(buffer).to_be_bytes()
+        } else {
+            [0xFF, 0xFF]
+        };
+        // These two bytes are always sent. They are either a valid CRC, or
+        // junk, depending on whether CRC mode was enabled.
+        self.write_bytes(&crc_bytes).await?;
+
+        let status = self.read_byte().await?;
+        if (status & DATA_RES_MASK) != DATA_RES_ACCEPTED {
+            Err(Error::WriteError)
+        } else {
+            Ok(())
+        }
+    }
+
+    fn cs_high(&mut self) -> Result<(), Error> {
+        self.cs.set_high().map_err(|_| Error::GpioError)
+    }
+
+    fn cs_low(&mut self) -> Result<(), Error> {
+        self.cs.set_low().map_err(|_| Error::GpioError)
+    }
+
+    /// Check the card is initialised.
+    async fn check_init(&mut self) -> Result<(), Error> {
+        if self.card_type.is_none() {
+            // If we don't know what the card type is, try and initialise the
+            // card. This will tell us what type of card it is.
+            self.acquire().await
+        } else {
+            Ok(())
+        }
+    }
+
+    /// Initializes the card into a known state (or at least tries to).
+    async fn acquire(&mut self) -> Result<(), Error> {
+        debug!("acquiring card with opts: {:?}", self.options);
+        let f = async |s: &mut Self| {
+            // Assume it hasn't worked
+            let mut card_type;
+            trace!("Reset card..");
+            // Supply minimum of 74 clock cycles without CS asserted.
+            s.cs_high()?;
+            s.write_bytes(&[0xFF; 10]).await?;
+            // Assert CS
+            s.cs_low()?;
+            // Enter SPI mode.
+            let mut delay = Delay::new(s.options.acquire_retries);
+            for _attempts in 1.. {
+                trace!("Enter SPI mode, attempt: {}..", _attempts);
+                match s.card_command(CMD0, 0).await {
+                    Err(Error::TimeoutCommand(0)) => {
+                        // Try again?
+                        warn!("Timed out, trying again..");
+                        // Try flushing the card as done here: https://github.com/greiman/SdFat/blob/master/src/SdCard/SdSpiCard.cpp#L170,
+                        // https://github.com/rust-embedded-community/embedded-sdmmc-rs/pull/65#issuecomment-1270709448
+                        for _ in 0..0xFF {
+                            s.write_byte(0xFF).await?;
+                        }
+                    }
+                    Err(e) => {
+                        return Err(e);
+                    }
+                    Ok(R1_IDLE_STATE) => {
+                        break;
+                    }
+                    Ok(_r) => {
+                        // Try again
+                        warn!("Got response: {:x}, trying again..", _r);
+                    }
+                }
+
+                delay
+                    .delay_async(&mut s.delayer, Error::CardNotFound)
+                    .await?;
+            }
+            // Enable CRC
+            debug!("Enable CRC: {}", s.options.use_crc);
+            // "The SPI interface is initialized in the CRC OFF mode in default"
+            // -- SD Part 1 Physical Layer Specification v9.00, Section 7.2.2 Bus Transfer Protection
+            if s.options.use_crc && s.card_command(CMD59, 1).await? != R1_IDLE_STATE {
+                return Err(Error::CantEnableCRC);
+            }
+            // Check card version
+            let mut delay = Delay::<Async>::new_command();
+            let arg = loop {
+                if s.card_command(CMD8, 0x1AA).await? == (R1_ILLEGAL_COMMAND | R1_IDLE_STATE) {
+                    card_type = CardType::SD1;
+                    break 0;
+                }
+                let mut buffer = [0xFF; 4];
+                s.transfer_bytes(&mut buffer).await?;
+                let status = buffer[3];
+                if status == 0xAA {
+                    card_type = CardType::SD2;
+                    break 0x4000_0000;
+                }
+                delay
+                    .delay_async(&mut s.delayer, Error::TimeoutCommand(CMD8))
+                    .await?;
+            };
+
+            let mut delay = Delay::<Async>::new_command();
+            while s.card_acmd(ACMD41, arg).await? != R1_READY_STATE {
+                delay
+                    .delay_async(&mut s.delayer, Error::TimeoutACommand(ACMD41))
+                    .await?;
+            }
+
+            if card_type == CardType::SD2 {
+                if s.card_command(CMD58, 0).await? != 0 {
+                    return Err(Error::Cmd58Error);
+                }
+                let mut buffer = [0xFF; 4];
+                s.transfer_bytes(&mut buffer).await?;
+                if (buffer[0] & 0xC0) == 0xC0 {
+                    card_type = CardType::SDHC;
+                }
+                // Ignore the other three bytes
+            }
+            debug!("Card version: {:?}", card_type);
+            s.card_type = Some(card_type);
+            Ok(())
+        };
+        let result = f(self).await;
+        self.cs_high()?;
+        let _ = self.read_byte();
+        result
+    }
+
+    /// Perform a function that might error with the chipselect low.
+    /// Always releases the chipselect, even if the function errors.
+    async fn with_chip_select<F, T>(&mut self, func: F) -> Result<T, Error>
+    where
+        F: AsyncFnOnce(&mut Self) -> Result<T, Error>,
+    {
+        self.cs_low()?;
+        let result = func(self).await;
+        self.cs_high()?;
+        result
+    }
+
+    /// Perform an application-specific command.
+    async fn card_acmd(&mut self, command: u8, arg: u32) -> Result<u8, Error> {
+        self.card_command(CMD55, 0).await?;
+        self.card_command(command, arg).await
+    }
+
+    /// Perform a command.
+    async fn card_command(&mut self, command: u8, arg: u32) -> Result<u8, Error> {
+        if command != CMD0 && command != CMD12 {
+            self.wait_not_busy(Delay::<Async>::new_command()).await?;
+        }
+
+        let mut buf = [
+            0x40 | command,
+            (arg >> 24) as u8,
+            (arg >> 16) as u8,
+            (arg >> 8) as u8,
+            arg as u8,
+            0,
+        ];
+        buf[5] = crc7(&buf[0..5]);
+
+        self.write_bytes(&buf).await?;
+
+        // skip stuff byte for stop read
+        if command == CMD12 {
+            let _result = self.read_byte().await?;
+        }
+
+        let mut delay = Delay::new_command();
+        loop {
+            let result = self.read_byte().await?;
+            if (result & 0x80) == ERROR_OK {
+                return Ok(result);
+            }
+            delay
+                .delay_async(&mut self.delayer, Error::TimeoutCommand(command))
+                .await?;
+        }
+    }
+
+    /// Receive a byte from the SPI bus by clocking out an 0xFF byte.
+    async fn read_byte(&mut self) -> Result<u8, Error> {
+        self.transfer_byte(0xFF).await
+    }
+
+    /// Send a byte over the SPI bus and ignore what comes back.
+    async fn write_byte(&mut self, out: u8) -> Result<(), Error> {
+        let _ = self.transfer_byte(out).await?;
+        Ok(())
+    }
+
+    /// Send one byte and receive one byte over the SPI bus.
+    async fn transfer_byte(&mut self, out: u8) -> Result<u8, Error> {
+        let mut read_buf = [0u8; 1];
+        self.spi
+            .transfer(&mut read_buf, &[out])
+            .await
+            .map_err(|_| Error::Transport)?;
+        Ok(read_buf[0])
+    }
+
+    /// Send multiple bytes and ignore what comes back over the SPI bus.
+    async fn write_bytes(&mut self, out: &[u8]) -> Result<(), Error> {
+        self.spi.write(out).await.map_err(|_e| Error::Transport)?;
+        Ok(())
+    }
+
+    /// Send multiple bytes and replace them with what comes back over the SPI bus.
+    async fn transfer_bytes(&mut self, in_out: &mut [u8]) -> Result<(), Error> {
+        self.spi
+            .transfer_in_place(in_out)
+            .await
+            .map_err(|_e| Error::Transport)?;
+        Ok(())
+    }
+
+    /// Spin until the card returns 0xFF, or we spin too many times and
+    /// timeout.
+    async fn wait_not_busy(&mut self, mut delay: Delay<Async>) -> Result<(), Error> {
+        loop {
+            let s = self.read_byte().await?;
+            if s == 0xFF {
+                break;
+            }
+            delay
+                .delay_async(&mut self.delayer, Error::TimeoutWaitNotBusy)
+                .await?;
+        }
+        Ok(())
+    }
+}
+
 /// Options for acquiring the card.
 #[cfg_attr(feature = "defmt-log", derive(defmt::Format))]
 #[derive(Debug)]
@@ -738,11 +1288,15 @@ pub enum CardType {
 ///
 /// Will let you call `delay` up to `max_retries` times before `delay` returns
 /// an error.
-struct Delay {
+struct Delay<M: Mode = Blocking> {
     retries_left: u32,
+    mode: PhantomData<M>,
 }
 
-impl Delay {
+impl<M> Delay<M>
+where
+    M: Mode,
+{
     /// The default number of retries for a read operation.
     ///
     /// At ~10us each this is ~100ms.
@@ -765,27 +1319,30 @@ impl Delay {
     pub const DEFAULT_COMMAND_RETRIES: u32 = 10_000;
 
     /// Create a new Delay object with the given maximum number of retries.
-    fn new(max_retries: u32) -> Delay {
+    fn new(max_retries: u32) -> Delay<M> {
         Delay {
             retries_left: max_retries,
+            mode: Default::default(),
         }
     }
 
     /// Create a new Delay object with the maximum number of retries for a read operation.
-    fn new_read() -> Delay {
+    fn new_read() -> Delay<M> {
         Delay::new(Self::DEFAULT_READ_RETRIES)
     }
 
     /// Create a new Delay object with the maximum number of retries for a write operation.
-    fn new_write() -> Delay {
+    fn new_write() -> Delay<M> {
         Delay::new(Self::DEFAULT_WRITE_RETRIES)
     }
 
     /// Create a new Delay object with the maximum number of retries for a command operation.
-    fn new_command() -> Delay {
+    fn new_command() -> Delay<M> {
         Delay::new(Self::DEFAULT_COMMAND_RETRIES)
     }
+}
 
+impl Delay<Blocking> {
     /// Wait for a while.
     ///
     /// Checks the retry counter first, and if we hit the max retry limit, the
@@ -799,6 +1356,26 @@ impl Delay {
             Err(err)
         } else {
             delayer.delay_us(10);
+            self.retries_left -= 1;
+            Ok(())
+        }
+    }
+}
+
+impl Delay<Async> {
+    /// Wait for a while.
+    ///
+    /// Checks the retry counter first, and if we hit the max retry limit, the
+    /// value `err` is returned. Otherwise we wait for 10us and then return
+    /// `Ok(())`.
+    async fn delay_async<T>(&mut self, delayer: &mut T, err: Error) -> Result<(), Error>
+    where
+        T: embedded_hal_async::delay::DelayNs,
+    {
+        if self.retries_left == 0 {
+            Err(err)
+        } else {
+            delayer.delay_us(10).await;
             self.retries_left -= 1;
             Ok(())
         }